data_intensive/chapter2/ddia_chat/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.3.2
alembic==1.14.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
faker==33.1.0